+++ packages/flow-mobile/flow-mobile-approval/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @coding-flow/flow-mobile-approval
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @coding-flow/flow-approval-presenter@0.1.4
+  - @coding-flow/flow-core@0.1.4
+  - @coding-flow/flow-icons@0.1.4
+  - @coding-flow/flow-mobile-form@0.1.4
+  - @coding-flow/flow-mobile-ui@0.1.4
+  - @coding-flow/flow-types@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes